MS-04: list folder created and moved files there
